[OSDF] Enable OSDF in ServiceMesh deployment
In case AAF is disabled the secret loading via AAF-SMS should not be called

Issue-ID: OPTFRA-1099

+++ osdf/adapters/aaf/sms.py
@@ -45,6 +45,9 @@ def preload_secrets():
     sms_url = config["aaf_sms_url"]
     timeout = config["aaf_sms_timeout"]
     cacert = config["aaf_ca_certs"]
+    if not sms_url:
+        debug_log.debug("SMS Disabled")
+        return
     sms_client = Client(url=sms_url, timeout=timeout, cacert=cacert)
     domain_uuid = sms_client.createDomain(domain)
     debug_log.debug(
@@ -64,46 +67,50 @@ def retrieve_secrets():
     timeout = config["aaf_sms_timeout"]
     cacert = config["aaf_ca_certs"]
     domain = config["secret_domain"]
-    sms_client = Client(url=sms_url, timeout=timeout, cacert=cacert)
-    secrets = sms_client.getSecretNames(domain)
-    for secret in secrets:
-        values = sms_client.getSecret(domain, secret)
-        secret_dict[secret] = values
-    debug_log.debug("Secret Dictionary Retrieval Success")
+    if sms_url:
+        sms_client = Client(url=sms_url, timeout=timeout, cacert=cacert)
+        secrets = sms_client.getSecretNames(domain)
+        for secret in secrets:
+            values = sms_client.getSecret(domain, secret)
+            secret_dict[secret] = values
+        debug_log.debug("Secret Dictionary Retrieval Success")
+    else:
+        debug_log.debug("SMS Disabled. Secrets not loaded")
     return secret_dict
 
 
 def load_secrets():
     config = osdf_config.deployment
     secret_dict = retrieve_secrets()
-    config['soUsername'] = secret_dict['so']['UserName']
-    config['soPassword'] = decrypt_pass(secret_dict['so']['Password'])
-    config['conductorUsername'] = secret_dict['conductor']['UserName']
-    config['conductorPassword'] = decrypt_pass(secret_dict['conductor']['Password'])
-    config['policyPlatformUsername'] = secret_dict['policyPlatform']['UserName']
-    config['policyPlatformPassword'] = decrypt_pass(secret_dict['policyPlatform']['Password'])
-    config['policyClientUsername'] = secret_dict['policyPlatform']['UserName']
-    config['policyClientPassword'] = decrypt_pass(secret_dict['policyPlatform']['Password'])
-    config['messageReaderAafUserId'] = secret_dict['dmaap']['UserName']
-    config['messageReaderAafPassword'] = decrypt_pass(secret_dict['dmaap']['Password'])
-    config['sdcUsername'] = secret_dict['sdc']['UserName']
-    config['sdcPassword'] = decrypt_pass(secret_dict['sdc']['Password'])
-    config['osdfPlacementUsername'] = secret_dict['osdfPlacement']['UserName']
-    config['osdfPlacementPassword'] = decrypt_pass(secret_dict['osdfPlacement']['Password'])
-    config['osdfPlacementSOUsername'] = secret_dict['osdfPlacementSO']['UserName']
-    config['osdfPlacementSOPassword'] = decrypt_pass(secret_dict['osdfPlacementSO']['Password'])
-    config['osdfPlacementVFCUsername'] = secret_dict['osdfPlacementVFC']['UserName']
-    config['osdfPlacementVFCPassword'] = decrypt_pass(secret_dict['osdfPlacementVFC']['Password'])
-    config['osdfCMSchedulerUsername'] = secret_dict['osdfCMScheduler']['UserName']
-    config['osdfCMSchedulerPassword'] = decrypt_pass(secret_dict['osdfCMScheduler']['Password'])
-    config['configDbUserName'] = secret_dict['configDb']['UserName']
-    config['configDbPassword'] = decrypt_pass(secret_dict['configDb']['Password'])
-    config['pciHMSUsername'] = secret_dict['pciHMS']['UserName']
-    config['pciHMSPassword'] = decrypt_pass(secret_dict['pciHMS']['Password'])
-    config['osdfPCIOptUsername'] = secret_dict['osdfPCIOpt']['UserName']
-    config['osdfPCIOptPassword'] = decrypt_pass(secret_dict['osdfPCIOpt']['Password'])
-    config['osdfOptEngineUsername'] = secret_dict['osdfOptEngine']['UserName']
-    config['osdfOptEnginePassword'] = decrypt_pass(secret_dict['osdfOptEngine']['Password'])
+    if secret_dict:
+        config['soUsername'] = secret_dict['so']['UserName']
+        config['soPassword'] = decrypt_pass(secret_dict['so']['Password'])
+        config['conductorUsername'] = secret_dict['conductor']['UserName']
+        config['conductorPassword'] = decrypt_pass(secret_dict['conductor']['Password'])
+        config['policyPlatformUsername'] = secret_dict['policyPlatform']['UserName']
+        config['policyPlatformPassword'] = decrypt_pass(secret_dict['policyPlatform']['Password'])
+        config['policyClientUsername'] = secret_dict['policyPlatform']['UserName']
+        config['policyClientPassword'] = decrypt_pass(secret_dict['policyPlatform']['Password'])
+        config['messageReaderAafUserId'] = secret_dict['dmaap']['UserName']
+        config['messageReaderAafPassword'] = decrypt_pass(secret_dict['dmaap']['Password'])
+        config['sdcUsername'] = secret_dict['sdc']['UserName']
+        config['sdcPassword'] = decrypt_pass(secret_dict['sdc']['Password'])
+        config['osdfPlacementUsername'] = secret_dict['osdfPlacement']['UserName']
+        config['osdfPlacementPassword'] = decrypt_pass(secret_dict['osdfPlacement']['Password'])
+        config['osdfPlacementSOUsername'] = secret_dict['osdfPlacementSO']['UserName']
+        config['osdfPlacementSOPassword'] = decrypt_pass(secret_dict['osdfPlacementSO']['Password'])
+        config['osdfPlacementVFCUsername'] = secret_dict['osdfPlacementVFC']['UserName']
+        config['osdfPlacementVFCPassword'] = decrypt_pass(secret_dict['osdfPlacementVFC']['Password'])
+        config['osdfCMSchedulerUsername'] = secret_dict['osdfCMScheduler']['UserName']
+        config['osdfCMSchedulerPassword'] = decrypt_pass(secret_dict['osdfCMScheduler']['Password'])
+        config['configDbUserName'] = secret_dict['configDb']['UserName']
+        config['configDbPassword'] = decrypt_pass(secret_dict['configDb']['Password'])
+        config['pciHMSUsername'] = secret_dict['pciHMS']['UserName']
+        config['pciHMSPassword'] = decrypt_pass(secret_dict['pciHMS']['Password'])
+        config['osdfPCIOptUsername'] = secret_dict['osdfPCIOpt']['UserName']
+        config['osdfPCIOptPassword'] = decrypt_pass(secret_dict['osdfPCIOpt']['Password'])
+        config['osdfOptEngineUsername'] = secret_dict['osdfOptEngine']['UserName']
+        config['osdfOptEnginePassword'] = decrypt_pass(secret_dict['osdfOptEngine']['Password'])
     cfg_base.http_basic_auth_credentials = creds.load_credentials(osdf_config)
     cfg_base.dmaap_creds = creds.dmaap_creds()
 
@@ -125,9 +132,12 @@ def delete_secrets():
     timeout = config["aaf_sms_timeout"]
     cacert = config["aaf_ca_certs"]
     domain = config["secret_domain"]
-    sms_client = Client(url=sms_url, timeout=timeout, cacert=cacert)
-    ret_val = sms_client.deleteDomain(domain)
-    debug_log.debug("Clean up complete")
+    if sms_url:
+        sms_client = Client(url=sms_url, timeout=timeout, cacert=cacert)
+        ret_val = sms_client.deleteDomain(domain)
+        debug_log.debug("Clean up complete")
+    else:
+        debug_log.debug("SMS Disabled. Secrets delete skipped")
     return ret_val
